// src/app/@shared/models/modif-mass-bals.model.ts
export interface ModifEnMassI{
        listeAdresses?: string[];
        etatBal?: string;
        filtreAdresse?: string;
        identifiantFrontal?: string;
        frontal?: string;
        seuil?: number;
        adressesIP?: string[];
        mailAlerte?: string;
        relaiMessagerie?: string     
}